+++ info.go
@@ -452,28 +452,3 @@ type infoResponseFull struct {
 	Info
 	SlackResponse
 }
-
-// GetBotByID is deprecated and returns nil
-func (info Info) GetBotByID(botID string) *Bot {
-	return nil
-}
-
-// GetUserByID is deprecated and returns nil
-func (info Info) GetUserByID(userID string) *User {
-	return nil
-}
-
-// GetChannelByID is deprecated and returns nil
-func (info Info) GetChannelByID(channelID string) *Channel {
-	return nil
-}
-
-// GetGroupByID is deprecated and returns nil
-func (info Info) GetGroupByID(groupID string) *Group {
-	return nil
-}
-
-// GetIMByID is deprecated and returns nil
-func (info Info) GetIMByID(imID string) *IM {
-	return nil
-}